adjust exports
