+++ package.json
@@ -3,7 +3,7 @@
   "name": "@digital-alchemy/type-writer",
   "repository": "https://github.com/Digital-Alchemy-TS/type-writer",
   "homepage": "https://docs.digital-alchemy.app/Type-Writer",
-  "version": "24.10.1",
+  "version": "24.10.2",
   "scripts": {
     "build": "rm -rf dist/; tsc",
     "lint": "eslint src",
@@ -40,20 +40,20 @@
   },
   "devDependencies": {
     "@cspell/eslint-plugin": "^8.15.4",
-    "@digital-alchemy/core": "^24.10.5",
+    "@digital-alchemy/core": "^24.10.8",
     "@digital-alchemy/hass": "^24.10.1",
-    "@eslint/compat": "^1.2.1",
+    "@eslint/compat": "^1.2.2",
     "@eslint/eslintrc": "^3.1.0",
     "@eslint/js": "^9.13.0",
     "@types/figlet": "^1.7.0",
-    "@types/jest": "^29.5.13",
+    "@types/jest": "^29.5.14",
     "@types/js-yaml": "^4.0.9",
-    "@types/node": "^22.7.7",
+    "@types/node": "^22.8.4",
     "@types/node-cron": "^3.0.11",
     "@types/validator": "^13.12.2",
     "@types/ws": "^8.5.12",
-    "@typescript-eslint/eslint-plugin": "8.10.0",
-    "@typescript-eslint/parser": "8.10.0",
+    "@typescript-eslint/eslint-plugin": "8.12.2",
+    "@typescript-eslint/parser": "8.12.2",
     "dayjs": "^1.11.13",
     "dotenv": "^16.4.5",
     "eslint": "9.13.0",
@@ -73,7 +73,7 @@
     "node-cron": "^3.0.3",
     "prettier": "^3.3.3",
     "ts-jest": "^29.2.5",
-    "tsx": "^4.19.1",
+    "tsx": "^4.19.2",
     "type-fest": "^4.26.1",
     "typescript": "^5.6.3",
     "ws": "^8.18.0"
